Fix tags and categories links when reading a post
Tags and categories links in article header were relative to current page,
which works well when people are on the homepage.
However, when you are in a blog post (/post/<…>), then, it's appended to it and
ends up as /post/<…>/tags/foo for instance, leading to a 404.
Add a / in front of links and be always relative to the baseurl of the website.